backend/lambda/users/requirements.txt
PyJWT
psycopg2-binary
bcrypt